requirements.txt
pyyaml==6.0
pandas==1.4.1
scikit-learn==1.0.2
tqdm==4.63.0
matplotlib==3.5.1
stockstats==0.4.1
ta==0.9.0
dcor==0.5.6
iisignature==0.24